Add `listLanguages` to access registered languages
Closes GH-17.

Co-authored-by: Titus Wormer <tituswormer@gmail.com>

This is a little hacky because Prism doesn't keep an explicit list of languages, and languages aren't
registered directly (a grammar is registered, which can arbitrarily modify the languages collection).
Instead, prism has an object that it uses to track its languages, but that object also has methods
attached to it, so its own properties include both languages and methods.