Fix double cleaning of HTML tags
In markdown, in HTML code (or XML, SVG, etc), a sanitation mechanism that Prism
needs was in place to drop what looks like HTML tags.
We don’t need that here in refractor, because we don’t have strings of HTML.
That is, we also don’t need to sanitize again, because with that in place we
dropped *safe* tags.

Closes GH-64.